game logic is moved to service class
