+++ game2048/components/Number/Number.module.scss
@@ -6,7 +6,6 @@
   border-radius: 5%;
   width: 15vmin;
   height: 15vmin;
-  // padding-left: px;
   color: black;
   animation: show 200ms ease-in-out;
   letter-spacing: 1px;
